Added the option for ctrl + u
When the user presses ctrl + u, the accessibility option should also pop up. I realize there is a bug on the settings menu, when pressing ctrl + u the accessibility checkbox doesn't change. I think I need to use a ref to fix that, but I till do it later because my brain is fried.